+++ ipysurfer/__init__.py
@@ -1,4 +1,4 @@
 """Google Summer of Code 3D rendering using ipyvolume."""
 
 from .io import read_activation_data, read_brain_mesh, read_morph
-from .viz import plot_brain_mesh
+from .viz import Brain
